feat: map ref custom hook
